Add "assignment" property to "DenyPolicyAssignmentResponseType"
The "DenyPolicyAssignmentResponseType" communicates back to the user one of
the three possible variants of "DenyPolicyAssignmentType" together with a
unique id assigned by the server. The specification says to destructure the
fields of the "DenyPolicyAssignmentType" into the response. In other words,
those fields which are found in the subclasses of "DenyPolicyAssignmentType"
must be placed directly into the "DenyPolicyAssignmentResponseType" instance.

This approach will work for dynamically typed languages but it definitely
does not work for the statically typed ones. In a statically typed language,
the compiler needs to know the fields of "DenyPolicyAssignmentResponseType"
at compile time, i.e. those fields must be specified. If those fields are not
specified, then the specification in its current form cannot be used for code
generation.

Adding an "assignment" property makes automated code generation possible.